Update to JMeter 3.0
JMeter 3.0 has been released, and ruby-jmeter has followed suit.

This requires jmeter 3.0, includes installation instructions for Mac,
and makes the script now agnostic to which minor version of jmeter is
installed in what spot on the machine.

Also adds a version constraint on the `ruby-jmeter` gem because the method
signature for `transactions_per_second` has changed: See
https://github.com/flood-io/ruby-jmeter/commit/95284fa80315948e44c9daf934f4a99eee0aa6fb