Tweak JSONPath expression and Cookie Manager for JMeter 3
So the previous JSONPath expression for getting the state ID of
New York did not work. I removed the `[0]` for gettting the first
piece of the array and instead rely on jmeter magic variables (`state_id_1`)
for getting the state ID.

Also, the Cookie Manager would not work with the `default` cookie
policy, and the `standard` cookie policy seems to be more like a
real browser.